backend/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
redis==4.2.2
selenium
openrouteservice
pdfplumber
faiss-cpu
bcrypt
scikit-learn
python-multipart